requirement.txt
streamlit
tensorflow
numpy
pillow
plotly
requests
